KNN Partially Fixed
In the last commit KNN did not run.
In this commit, I've touched it up enough to run without errors, the algorithm is not efficient, the load time on game startup is significant.
It needs some efficiency upgrades and adjustments to its selection of outputs: it is currently not very accurate.