Major Update!
Fixed bug where the size of each player's "brain" of networks was not being reset.
Added an implementation to store various network configurations separately. Which allows for experimentation of network topology and a record of those experiments.
Generation cards are now time stamped.

Additionally, we added a python script which processes the shot history card by removing duplicates and splitting the data into 42 sub files, (for 21 wind conditions for each player)

Co-Authored-By: Carl Hartry Jr. <71412408+CHartryJr@users.noreply.github.com>
Co-Authored-By: Ser3ntis <58407045+Ser3ntis@users.noreply.github.com>